refactor: split mailer service into focused modules
Break the single mailer.ts file into mailer/ with a public
sendClubInviteCodeEmail entrypoint, a private Resend transport, and
separate HTML/text templates, so send logic and email markup no
longer live in one file. Drop internal validation/error classes and
the top-level services barrel export, none of which had real
consumers beyond this module.
