+++ src/server/api/modules/invitations/usecases/generate-invitation-code.usecase.ts
@@ -6,13 +6,9 @@ import type {
   GenerateInvitationCodeInputDTO,
   GenerateInvitationCodeOutputDTO,
 } from "@/server/api/modules/invitations/dto";
-import {
-  sendClubInviteCodeEmail,
-  MailValidationError,
-  MailDeliveryError,
-} from "@/server/services/mailer";
+import { sendClubInviteCodeEmail } from "@/server/services/mailer";
 
-// Client-ratified TTL — see the contract comment at the top of src/server/services/mailer.ts.
+// Client-ratified TTL — see the contract comment at the top of src/server/services/mailer/mailer.ts.
 const INVITE_CODE_TTL_DAYS = 14;
 
 const ALPHABET = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
@@ -46,27 +42,16 @@ export const generateInvitationCode = async (
     await sendClubInviteCodeEmail({
       to: invitationCode.email,
       inviteCode: invitationCode.inviteCode,
-      clubName: input.clubName,
     });
   } catch (err) {
     // The DB row is already committed and stays that way — deleting it here would silently
     // discard a real invalidation of whatever code preceded it. A retried `generate` call is
     // safe: it invalidates this row too and mints a fresh code, so the admin can just retry.
-    if (err instanceof MailValidationError) {
-      throw new TRPCError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: "The recipient email was rejected by the mail provider. Please retry.",
-        cause: err,
-      });
-    }
-    if (err instanceof MailDeliveryError) {
-      throw new TRPCError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: "The code was generated, but the email could not be sent. Please retry.",
-        cause: err,
-      });
-    }
-    throw err;
+    throw new TRPCError({
+      code: "INTERNAL_SERVER_ERROR",
+      message: "The code was generated, but the email could not be sent. Please retry.",
+      cause: err,
+    });
   }
 
   return invitationCode.toDTO();
