--- src/server/services/index.ts
@@ -1,2 +1,0 @@
-export * from "@/server/services/mailer";
-export * from "@/server/services/r2";
