--- src/server/services/mailer.ts
@@ -1,345 +1,0 @@
-import { Resend } from "resend";
-import { z } from "zod";
-import { env } from "@/config/env";
-
-/**
- * CONTRACT REQUIREMENTS HANDED OFF TO #69 / #108
- *
- * None of the following can be enforced by code inside this module (#68) — they are written
- * down here so the requirement is not silently assumed by whoever builds the consumer.
- *
- * (a) GET /register MUST be non-consuming [#108]. Email security gateways (Microsoft Defender
- *     Safe Links, Proofpoint, Mimecast) auto-GET every URL in an inbound message within seconds
- *     of delivery, before any human opens it. If validation consumes on GET, the scanner burns
- *     the single-use code and the real recipient sees "already used". The code may be READ from
- *     the query string to prefill the form; it must only be checked and consumed on the
- *     subsequent POST.
- * (b) Invite codes MUST carry a server-side expiry [#69 issues it, #108 checks it].
- *     TTL is 14 days (client-ratified). Expiry is a property of the code, never of the send: an
- *     expired code is rejected regardless of whether the email was ever successfully delivered.
- *     Re-triggering a failed send with the SAME code must NOT extend expiresAt, or repeated
- *     retries silently mint an immortal credential. If already expired, issue a fresh code.
- * (c) OPS: Resend open-and-click tracking MUST be disabled for the sending domain. Resend link
- *     tracking rewrites every href into a Resend redirect URL, so the code would transit a
- *     third party and every gateway prefetch would be logged there. It is a per-domain
- *     dashboard setting, not a per-call API flag, so no code change can guarantee it. Requires
- *     human verification, stated in the PR description.
- * (d) /register MUST send Referrer-Policy: no-referrer or same-origin [#108]. next.config.js
- *     currently defines no headers() function at all, so the default applies. Without it, any
- *     third-party resource the register page loads receives the full referring URL —
- *     including the invite code — in the Referer header.
- *
- * RESIDUAL RISK ACCEPTED: even with (a)-(d), a code in a query string persists in browser
- * history, in Vercel/edge/CDN access logs (which capture full request lines including query
- * strings), and in intermediate proxy logs. The expiry in (b) is the primary mitigation for
- * that persistence.
- */
-
-const resend = new Resend(env.RESEND_API_KEY);
-
-const PROVIDER_NAME = "resend";
-const SUBJECT = "รหัสเชิญเข้าร่วม CUatClub ของคุณ";
-const REGISTER_PATH = "/register";
-const INVITE_CODE_PARAM = "inviteCode";
-const LOGO_PATH = "/images/logo.png";
-const FONT_STACK = "'IBM Plex Sans Thai', 'Sarabun', Tahoma, Arial, sans-serif";
-
-export class MailValidationError extends Error {
-  readonly field: "to" | "inviteCode" | "clubName";
-
-  constructor(field: "to" | "inviteCode" | "clubName", message: string) {
-    super(message);
-    this.name = "MailValidationError";
-    this.field = field;
-  }
-}
-
-export class MailDeliveryError extends Error {
-  readonly retryable: boolean;
-  readonly providerName: string;
-
-  constructor(message: string, retryable: boolean, providerName: string) {
-    super(message);
-    this.name = "MailDeliveryError";
-    this.retryable = retryable;
-    this.providerName = providerName;
-  }
-}
-
-export interface SendClubInviteCodeEmailParams {
-  /** Club email; single recipient by design, never an array. */
-  to: string;
-  /** Generated by #69, not here. */
-  inviteCode: string;
-  /** Optional, used in the greeting. */
-  clubName?: string;
-}
-
-export interface SendClubInviteCodeEmailResult {
-  messageId: string;
-}
-
-/**
- * Zod .email() is a shape check, not a single-address check. Do not assume it rejects
- * "a@x.com,b@y.com" or embedded CRLF. These refinements make the single-recipient
- * anti-fan-out invariant an explicit, tested property of this file.
- */
-const recipientSchema = z
-  .string()
-  .min(1)
-  .max(254)
-  .email()
-  .refine((v) => !/[,;]/.test(v), { message: "must be a single email address" })
-  .refine((v) => !/[\r\n\t]/.test(v), { message: "must not contain control characters" })
-  .refine((v) => (v.match(/@/g) ?? []).length === 1, { message: "must contain exactly one @" })
-  .refine((v) => v === v.trim(), { message: "must not have surrounding whitespace" });
-
-const inviteCodeSchema = z
-  .string()
-  .min(1)
-  .max(64)
-  .regex(/^[A-Za-z0-9_-]+$/, { message: "invalid invite code format" });
-
-const clubNameSchema = z
-  .string()
-  .min(1)
-  .max(120)
-  .refine((v) => !/[\r\n]/.test(v), { message: "must not contain line breaks" })
-  .optional();
-
-function escapeHtml(value: string): string {
-  return value
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-function maskEmail(email: string): string {
-  const atIndex = email.indexOf("@");
-  if (atIndex <= 0) return "***";
-  const domain = email.slice(atIndex);
-  return `${email[0]}***${domain}`;
-}
-
-function buildRegisterUrl(inviteCode: string): string {
-  const url = new URL(REGISTER_PATH, env.APP_BASE_URL);
-  url.searchParams.set(INVITE_CODE_PARAM, inviteCode);
-  return url.toString();
-}
-
-function buildLogoUrl(): string {
-  return new URL(LOGO_PATH, env.APP_BASE_URL).toString();
-}
-
-function renderHtml(inviteCodeHtml: string, registerUrlHtml: string, logoUrlHtml: string) {
-  return `<!DOCTYPE html>
-<html lang="th" xmlns="http://www.w3.org/1999/xhtml">
-<head>
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
-<title>รหัสเชิญเข้าร่วม CUatClub ของคุณ</title>
-</head>
-<body style="margin:0; padding:0; background-color:#f0f0f0; -webkit-text-size-adjust:100%; -ms-text-size-adjust:100%;">
-
-  <!-- Preheader (hidden, shows in inbox preview) -->
-  <div style="display:none; max-height:0; max-width:0; overflow:hidden; opacity:0; mso-hide:all; font-size:1px; line-height:1px; color:#f0f0f0;">
-    รหัสเชิญเข้าร่วม CUatClub ของคุณอยู่ในอีเมลนี้ ใช้งานได้เพียงครั้งเดียว
-  </div>
-  <!-- Spacer to push preheader text away from visible content in some clients -->
-  <div style="display:none; max-height:0; overflow:hidden; mso-hide:all;">
-    &nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;&nbsp;&zwnj;
-  </div>
-
-  <table role="presentation" width="100%" cellpadding="0" cellspacing="0" border="0" style="background-color:#f0f0f0;">
-    <tr>
-      <td align="center" style="padding:24px 16px;">
-
-        <table role="presentation" width="600" cellpadding="0" cellspacing="0" border="0" style="width:600px; max-width:600px; background-color:#ffffff;">
-
-          <!-- Body -->
-          <tr>
-            <td style="padding:40px 40px 8px 40px; font-family:${FONT_STACK};">
-              <img src="${logoUrlHtml}" width="91" height="48" alt="CUatClub" style="display:block; margin:0 0 20px 0; border:0; outline:none;" />
-              <h1 style="margin:0; font-size:20px; line-height:32px; font-weight:700; color:#0a0a0a;">นี่คือรหัสเชิญเข้าร่วมของคุณ</h1>
-              <p style="margin:12px 0 0 0; font-size:14px; line-height:23px; color:#595959;">
-                กรอกรหัสนี้บนหน้าลงทะเบียนเพื่อเริ่มใช้งานบัญชีของคุณ
-              </p>
-            </td>
-          </tr>
-
-          <!-- Code block -->
-          <tr>
-            <td align="center" style="padding:24px 40px 8px 40px;">
-              <table role="presentation" cellpadding="0" cellspacing="0" border="0">
-                <tr>
-                  <td align="center" style="background-color:rgba(248,222,232,0.5); border-radius:6px; padding:18px 64px;">
-                    <span style="font-family:${FONT_STACK}; font-size:32px; line-height:38px; font-weight:700; color:#dd598c;">${inviteCodeHtml}</span>
-                  </td>
-                </tr>
-              </table>
-            </td>
-          </tr>
-
-          <!-- Instructions -->
-          <tr>
-            <td align="center" style="padding:16px 40px 0 40px; font-family:${FONT_STACK};">
-              <p style="margin:0; font-size:14px; line-height:23px; color:#595959;">
-                รหัสนี้หมดอายุใน 14 วัน และใช้งานได้เพียงครั้งเดียว
-              </p>
-            </td>
-          </tr>
-
-          <!-- CTA button -->
-          <tr>
-            <td align="center" style="padding:28px 40px 8px 40px;">
-              <table role="presentation" cellpadding="0" cellspacing="0" border="0">
-                <tr>
-                  <td align="center" style="background-color:#dd598c; border-radius:8px;">
-                    <a href="${registerUrlHtml}" target="_blank" style="display:inline-block; padding:12px 32px; font-family:${FONT_STACK}; font-size:15px; line-height:20px; font-weight:600; color:#ffffff; text-decoration:none; border-radius:8px;">ลงทะเบียนชมรม</a>
-                  </td>
-                </tr>
-              </table>
-            </td>
-          </tr>
-
-          <!-- Fallback URL -->
-          <tr>
-            <td align="center" style="padding:8px 40px 32px 40px; font-family:${FONT_STACK};">
-              <p style="margin:0; font-size:12px; line-height:18px; color:#707070; word-break:break-all;">
-                หรือคัดลอกลิงก์นี้ไปวางในเบราว์เซอร์:<br>
-                <a href="${registerUrlHtml}" target="_blank" style="color:#dd598c; text-decoration:underline;">${registerUrlHtml}</a>
-              </p>
-            </td>
-          </tr>
-
-          <!-- Footer -->
-          <tr>
-            <td style="padding:20px 40px; border-top:1px solid #ececec; font-family:${FONT_STACK};" align="center">
-              <p style="margin:0; font-size:12px; line-height:18px; color:#707070;">
-                ไม่ได้ขอรหัสเชิญเข้าร่วม CUatClub ใช่หรือไม่? คุณสามารถละเว้นอีเมลฉบับนี้ได้อย่างปลอดภัย รหัสนี้จะไม่สามารถใช้งานได้กับผู้อื่น
-              </p>
-              <p style="margin:8px 0 0 0; font-size:12px; line-height:18px; color:#acacac;">
-                &copy; CUatClub
-              </p>
-            </td>
-          </tr>
-
-        </table>
-
-      </td>
-    </tr>
-  </table>
-
-</body>
-</html>`;
-}
-
-function renderText(inviteCode: string, registerUrl: string): string {
-  return `นี่คือรหัสเชิญเข้าร่วม CUatClub ของคุณ: ${inviteCode}
-
-กรอกรหัสนี้บนหน้าลงทะเบียนเพื่อเริ่มใช้งานบัญชีของคุณ รหัสนี้หมดอายุใน 14 วัน และใช้งานได้เพียงครั้งเดียว
-
-ลงทะเบียนชมรม: ${registerUrl}
-
-ไม่ได้ขอรหัสเชิญเข้าร่วม CUatClub ใช่หรือไม่? คุณสามารถละเว้นอีเมลฉบับนี้ได้อย่างปลอดภัย รหัสนี้จะไม่สามารถใช้งานได้กับผู้อื่น
-
-CUatClub`;
-}
-
-/**
- * Send a club's single-use invitation code by email.
- *
- * Authorization obligations for the caller (#69):
- * - The endpoint that triggers this send MUST be an `adminProcedure`.
- * - The usecase MUST resolve `to` from a persisted club record rather than accepting a
- *   caller-supplied recipient string.
- */
-export async function sendClubInviteCodeEmail(
-  params: SendClubInviteCodeEmailParams
-): Promise<SendClubInviteCodeEmailResult> {
-  const toResult = recipientSchema.safeParse(params.to);
-  if (!toResult.success) {
-    throw new MailValidationError("to", "Invalid recipient email address.");
-  }
-
-  const inviteCodeResult = inviteCodeSchema.safeParse(params.inviteCode);
-  if (!inviteCodeResult.success) {
-    throw new MailValidationError("inviteCode", "Invalid invite code.");
-  }
-
-  // clubName is validated but not currently rendered anywhere in the template (design has no
-  // greeting line) — kept as a public param since callers may already pass it, and future
-  // copy changes may want it again.
-  const clubNameResult = clubNameSchema.safeParse(params.clubName);
-  if (!clubNameResult.success) {
-    throw new MailValidationError("clubName", "Invalid club name.");
-  }
-
-  const to = toResult.data;
-  const inviteCode = inviteCodeResult.data;
-
-  const registerUrl = buildRegisterUrl(inviteCode);
-  const registerUrlHtml = escapeHtml(registerUrl);
-  const logoUrlHtml = escapeHtml(buildLogoUrl());
-
-  const html = renderHtml(escapeHtml(inviteCode), registerUrlHtml, logoUrlHtml);
-  // The plain-text part is deliberately NOT escaped — HTML entities in a text/plain body are
-  // a rendering bug, not a security control.
-  const text = renderText(inviteCode, registerUrl);
-
-  try {
-    const res = await resend.emails.send({
-      from: `CUatClub <${env.EMAIL_FROM}>`,
-      to,
-      subject: SUBJECT,
-      html,
-      text,
-    });
-
-    if (res.error) {
-      const statusCode = res.error.statusCode;
-      const retryable = statusCode == null || statusCode === 429 || statusCode >= 500;
-      console.error("mailer: resend delivery failed", {
-        name: res.error.name,
-        statusCode: res.error.statusCode,
-        recipient: maskEmail(to),
-      });
-      throw new MailDeliveryError(
-        "Unable to send email right now. Please try again later.",
-        retryable,
-        PROVIDER_NAME
-      );
-    }
-
-    if (!res.data?.id) {
-      console.error("mailer: resend returned no message id", {
-        recipient: maskEmail(to),
-      });
-      throw new MailDeliveryError(
-        "Unable to send email right now. Please try again later.",
-        false,
-        PROVIDER_NAME
-      );
-    }
-
-    return { messageId: res.data.id };
-  } catch (err) {
-    if (err instanceof MailDeliveryError) throw err;
-
-    const name = err instanceof Error ? err.name : undefined;
-    const message = err instanceof Error ? err.message : undefined;
-    console.error("mailer: resend send threw", {
-      name,
-      message,
-      recipient: maskEmail(to),
-    });
-    throw new MailDeliveryError(
-      "Unable to send email right now. Please try again later.",
-      true,
-      PROVIDER_NAME
-    );
-  }
-}
