+++ src/server/services/mailer/index.ts
@@ -1,0 +1,1 @@
+export { sendClubInviteCodeEmail } from "./mailer";
