+++ src/server/services/mailer/templates/index.ts
@@ -1,0 +1,2 @@
+export { renderClubInviteCodeEmailHtml } from "./club-invite-code-email.html";
+export { renderClubInviteCodeEmailText } from "./club-invite-code-email.text";
